commit many 12/15_2

+++ SeaHours/SeaHours/Character.swift
@@ -46,7 +46,7 @@ class Character {
         }
     }
 
-    func damageCalculate(skillDamage : Int , enemyDefence : Int) -> Int {
+    func damageCalculate(skillDamage : Double , enemyDefence : Int) -> Int {
         //最終的なダメージ　- 敵の防御力
         var damege : Int
         var randomNumber1 : Double
@@ -57,7 +57,7 @@ class Character {
         randomNumber1 = 1 + randomNumber1
         randomNumber2 = 1 + randomNumber2
         
-        damege = Int(Double(skillDamage) * randomNumber1 - Double(enemyDefence) * randomNumber2)
+        damege = Int(skillDamage * randomNumber1 - Double(enemyDefence) * randomNumber2)
         if (damege < 0){
             damege = 0
         }
